Add class, initialize module, include module Create class MyList Add file my_enumerable Include enumerable in the class Update README.md
