+++ README.md
@@ -1,1 +1,51 @@
-# enumerable
+# Enumerable
+
+- This repository shows you how to use a module inside a class. There is MyList class and MyEnumerable module. The Enumerable module implements a subset of the functionality of Enumerable.
+
+## Built With
+
+- Major languages (Ruby)
+
+**Getting Started**
+To get a local copy up and running follow these simple example steps.
+
+**Prerequisites**
+
+- A text editor(preferably Visual Studio Code)
+
+**Install**
+
+- Ruby
+
+**Using it Locally**
+
+- Clone the project
+  `git clone git@github.com:thecodechaser/my-ruby-enumerable.git`
+  `cd my-ruby-enumerable`
+
+**To Start the programs**
+`open terminal`
+`run ruby my_list.rb`
+
+## Authors
+
+👤 **Yasin Warsame**
+
+- GitHub: [@Yazino12](https://github.com/Yazino12)
+
+👤 **Christian Biringanine**
+
+- GitHub: [@christianbiring1](https://github.com/christianbiring1)
+
+## 🤝 Contributing
+
+- Contributions, issues, and feature requests are welcome!
+- Feel free to check the issues page.
+
+## Show your support
+
+- Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Inspiration: Microverse
