+++ my_enumerable.rb
@@ -1,0 +1,2 @@
+module MyEnumerable
+end
